Add FfiGen class
