+++ pkgs/ffigen/lib/ffigen.dart
@@ -9,5 +9,5 @@ library ffigen;
 
 export 'src/code_generator.dart' show Library;
 export 'src/config_provider.dart' show Config, YamlConfig;
-export 'src/generator.dart' show generate, setupLogger;
+export 'src/ffigen.dart';
 export 'src/header_parser.dart' show parse;
